+++ tpccJava/src/main/java/org/example/DoPaymentOutput.java
@@ -30,7 +30,7 @@ public class DoPaymentOutput {
     public String c_since;
     public String c_credit;
     public int c_credit_lim;
-    public int c_discount;
+    public Double c_discount;
     public Double c_balance;
     public String c_data;
